+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Fix issue [#104](https://github.com/intersystems/language-server/issues/104): Strip out "style" HTML tags from class reference documentation
 - Fix issue [#106](https://github.com/intersystems/language-server/issues/106): #dim provides no code completion inside a trigger
 - Fix issue [#107](https://github.com/intersystems/language-server/issues/107): Add "Go To Declaration" support
+- Fix issue [#108](https://github.com/intersystems/language-server/issues/108): Add "Go To Declaration" support for variables in the PublicList
 
 ## [1.0.5] - 2020-11-12
 - Fix issue [#52](https://github.com/intersystems/language-server/issues/52): Request textDocument/documentSymbol failed with message: Cannot read property 'p' of undefined